+++ compiler/src/utils/constants.ts
@@ -1,0 +1,30 @@
+/**
+ * The prefix for internal variables inside the compiler output
+ */
+export const internalPrefix = "&";
+export const discardedName = `${internalPrefix}_`;
+
+export const itemNames = [
+  "copper",
+  "lead",
+  "metaglass",
+  "graphite",
+  "sand",
+  "coal",
+  "titanium",
+  "thorium",
+  "scrap",
+  "silicon",
+  "plastanium",
+  "phaseFabric",
+  "surgeAlloy",
+  "sporePod",
+  "blastCompound",
+  "pyratite",
+  "beryllium",
+  "tungsten",
+  "oxide",
+  "carbide",
+  "fissileMatter",
+  "dormantCyst",
+];
